pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.12",
    "python-dotenv>=1.2.1",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W"]
